feat: MES sea-orm 实体重构 + API 对齐 + Docker 部署
- 所有 63 张表的实体文件迁移至 sea-orm DeriveEntityModel 格式
- DAO 层批量修复 RelationTrait、InsertResult、Claims 等 API 适配
- API 层 (auth/equipment/quality/schedule/warehouse) 对齐新实体
- model 层数据结构同步更新 (equipment/master_data/production/quality/schedule/system/warehouse)
- MQTT handler 和 JWT utils 适配变更
- 新增 Dockerfile 支持容器化部署 (backend + frontend)
- 编译通过，服务正常运行

+++ mes-backend/src/db/dao/equipment_dao.rs
@@ -7,7 +7,7 @@ use sea_orm::{
 
 #[derive(Debug)]
 pub struct EquipmentFilter {
-    pub status: Option<i16>,
+    pub status: Option<i32>,
 }
 
 pub async fn list(
@@ -96,7 +96,7 @@ pub async fn upsert_mqtt_config(
 pub async fn append_status_log(
     conn: ConnRef<'_>,
     equipment_id: i32,
-    status: i16,
+    status: i32,
     running_param: Option<serde_json::Value>,
     error_code: Option<String>,
     error_desc: Option<String>,
